Removed activites package moved splashscreen to ui package
