+++ app/src/main/AndroidManifest.xml
@@ -49,7 +49,7 @@
         </provider>
 
         <activity
-            android:name=".activities.SplashScreenActivity"
+            android:name=".ui.splashScreen.SplashScreenActivity"
             android:exported="true"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar">
             <intent-filter>
