--- app/src/main/java/org/wit/recipes/activities/MapActivity.kt
@@ -1,34 +1,0 @@
-package org.wit.recipes.activities
-
-import androidx.appcompat.app.AppCompatActivity
-import android.os.Bundle
-
-import com.google.android.gms.maps.CameraUpdateFactory
-import com.google.android.gms.maps.GoogleMap
-import com.google.android.gms.maps.OnMapReadyCallback
-import com.google.android.gms.maps.SupportMapFragment
-import com.google.android.gms.maps.model.LatLng
-import com.google.android.gms.maps.model.MapStyleOptions
-import com.google.android.gms.maps.model.MarkerOptions
-import org.wit.recipes.R
-
-class MapActivity : AppCompatActivity(), OnMapReadyCallback {
-
-    private lateinit var mMap: GoogleMap
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_map)
-        val mapFragment = supportFragmentManager
-            .findFragmentById(R.id.map) as SupportMapFragment
-        mapFragment.getMapAsync(this)
-    }
-
-    override fun onMapReady(googleMap: GoogleMap) {
-        mMap = googleMap
-        val waterford = LatLng(52.257298, -7.111579)
-        mMap.addMarker(MarkerOptions().position(waterford).title("Marker in Waterford"))
-        mMap.setMapStyle(MapStyleOptions(resources.getString(R.string.style_json)));
-        mMap.moveCamera(CameraUpdateFactory.newLatLngZoom(waterford, 15f))
-    }
-}
