+++ app/src/main/res/layout/activity_splash_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorAccent"
-    tools:context=".activities.SplashScreenActivity">
+    tools:context=".ui.splashScreen.SplashScreenActivity">
 
 
     <TextView
